# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/switch-workspace": {
      "post": {
        "operationId": "AuthController_switchWorkspace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace switched successfully"
          },
          "403": {
            "description": "Access denied to target workspace"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Switch active workspace context",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid or expired verification code"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Verify email address using 6-digit OTP code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code sent successfully"
          },
          "400": {
            "description": "Email is already verified"
          },
          "404": {
            "description": "User not found"
          },
          "429": {
            "description": "Cooldown active (please wait 60 seconds)"
          }
        },
        "summary": "Resend email verification 6-digit OTP code",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Current password is incorrect"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password from dashboard settings (authenticated)",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kvik Backend API",
    "description": "AI SaaS Sales Agent for Kazakhstan and CIS Market",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "SwitchWorkspaceDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {}
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {}
      },
...
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}